Added all the project requirements except Form Validation for all entities except Customer

+++ TravelItinerary/Shared/Domain/Payment.cs
@@ -10,9 +10,15 @@ namespace TravelItinerary.Shared.Domain
     {
         public string PaymentType { get; set; }
         public DateTime PaymentDate { get; set; }
-        public int PaymentAmount { get; set; }
+        public double PaymentAmount { get; set; }
         public string PaymentDescription { get; set; }
         public int CustomerId { get; set; }
         public virtual Customer Customer { get; set; }
+        //public int AccommodationId { get; set; }
+        //public virtual Accommodation Accommodation { get; set; }
+        //public int FlightId { get; set; }
+        //public virtual Flight Flight { get; set; }
+        //public int FligAccoId { get; set; }
+        //public virtual FligAcco FligAcco { get; set; }
     }
 }
